# swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "0.0.1"
  title: "Swagger Flights API"
host: "localhost:3000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /cities:
    post:
      tags:
      - "cities"
      summary: "Adiciona um novo cities"
      description: ""
      operationId: "addCity"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Dados do cities a ser salvo"
        required: true
        schema:
          $ref: "#/definitions/cities"
      responses:
        "200":
          description: "SUCCESS"
        "500":
          description: "INTERNAL SERVER ERROR"
    get:
      tags:
      - "cities"
      summary: "Retorna todos os voos"
      description: ""
      operationId: "getCities"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "SUCCESS"
        "500":
          description: "INTERNAL SERVER ERROR"      
  /cities/{cityId}:        
    put:
      tags:
      - "cities"
      summary: "Altera um voo existente"
      description: ""
      operationId: "updateCity"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "cityId"
        in: "path"
        description: "Id da cidade"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Dados que precisam ser alterados no cities selecionado"
        required: true
        schema:
          $ref: "#/definitions/cities"
      responses:
        "200":
          description: "SUCCESS"
        "500":
          description: "INTERNAL SERVER ERROR"
    delete:
      tags:
      - "cities"
      summary: "Deleta um cities existente"
      description: ""
      operationId: "deleteCity"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "cityId"
        in: "path"
        description: "Id da cidade"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "SUCCESS"
        "500":
          description: "INTERNAL SERVER ERROR"
    get:
      tags:
      - "cities"
      summary: "Retorna um cities especifico"
      description: ""
      operationId: "getCity"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "cityId"
        in: "path"
        description: "Id da cidade"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "SUCCESS"
        "500":
          description: "INTERNAL SERVER ERROR"
...
definitions:
...
  cities:
    type: "object"
    properties:
      country:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "cities"
...